Improved the section on Linux build instructions
Improved build instructions especially for Debian/Ubuntu. Made it less intimidating, moved the info which is for obsolete systems to a troubleshooting section, added the new Debian packages which can be used for pushing out new features for testing and also added information about the master branch with info about dynamic testnet support.